+++ docs/tailwind.config.cjs
@@ -22,9 +22,17 @@ module.exports = {
   ],
   ...getExpoTheme(
     {
+      screens: {
+        '2xl': '1500px',
+        '2xl-gutters': '1572px',
+      },
       backgroundImage: theme => ({
         'default-fade': `linear-gradient(to bottom, ${theme('backgroundColor.default')}, transparent)`,
         'default-fade-down': `linear-gradient(to bottom, transparent, ${theme('backgroundColor.default')})`,
+        'cell-quickstart-fade': `linear-gradient(${theme('backgroundColor.subtle')} 15%, #21262d00 100%)`,
+        'cell-quickstart-pattern': "url('/static/images/home/QuickStartPattern.svg')",
+        'cell-tutorial-fade': `linear-gradient(var(--blue4) 15%, #201d5200 100%)`,
+        'cell-tutorial-pattern': "url('/static/images/home/TutorialPattern.svg')",
         appjs: "url('/static/images/appjs.svg'), linear-gradient(#0033cc, #0033cc)",
       }),
       fontSize: {