+++ packages/@expo/cli/ts-declarations/metro-babel-transformer/index.d.ts
@@ -39,6 +39,7 @@ declare module 'metro-babel-transformer' {
     transform: (args: BabelTransformerArgs) => {
       ast: Ast;
       metadata?: BabelFileMetadata & {
+        reactClientReference?: string;
         metro?: {
           functionMap?: FBSourceFunctionMap | null;
         };