+++ packages/@expo/metro-config/src/transform-worker/metro-transform-worker.ts
@@ -61,6 +61,7 @@ interface JSFile extends BaseFile {
   readonly ast?: ParseResult | null;
   readonly type: JSFileType;
   readonly functionMap: FBSourceFunctionMap | null;
+  readonly reactClientReference?: string;
 }
 
 interface JSONFile extends BaseFile {
@@ -75,8 +76,14 @@ interface TransformationContext {
 
 interface TransformResponse {
   readonly dependencies: readonly TransformResultDependency[];
-  readonly output: readonly JsOutput[];
+  readonly output: readonly ExpoJsOutput[];
 }
+
+export type ExpoJsOutput = Pick<JsOutput, 'type'> & {
+  readonly data: JsOutput['data'] & {
+    readonly reactClientReference?: string;
+  };
+};
 
 // asserts non-null
 function nullthrows<T extends object>(x: T | null, message?: string): NonNullable<T> {
@@ -169,6 +176,9 @@ async function transformJS(
   file: JSFile,
   { config, options, projectRoot }: TransformationContext
 ): Promise<TransformResponse> {
+  // const targetEnv = options.customTransformOptions?.environment;
+  // const isServerEnv = targetEnv === 'node' || targetEnv === 'react-server';
+
   // Transformers can output null ASTs (if they ignore the file). In that case
   // we need to parse the module source code to get their AST.
   let ast: babylon.ParseResult<types.File> =
@@ -313,6 +323,10 @@ async function transformJS(
         allowOptionalDependencies: config.allowOptionalDependencies,
         dependencyMapName: config.unstable_dependencyMapReservedName,
         unstable_allowRequireContext: config.unstable_allowRequireContext,
+
+        // NOTE(EvanBacon): Allow arbitrary imports in server environments.
+        // This requires a patch to Metro collectDeps.
+        // allowArbitraryImport: isServerEnv,
       };
 
       ({ ast, dependencies, dependencyMapName } = collectDependencies(ast, opts));
@@ -390,13 +404,14 @@ async function transformJS(
     ));
   }
 
-  const output: JsOutput[] = [
+  const output: ExpoJsOutput[] = [
     {
       data: {
         code,
         lineCount: countLines(code),
         map,
         functionMap: file.functionMap,
+        reactClientReference: file.reactClientReference,
       },
       type: file.type,
     },
@@ -427,6 +442,7 @@ async function transformAsset(
     type: 'js/module/asset',
     ast: result.ast,
     functionMap: null,
+    reactClientReference: result.reactClientReference,
   };
 
   return transformJS(jsFile, context);
@@ -467,6 +483,7 @@ async function transformJSWithBabel(
       // Fallback to deprecated explicitly-generated `functionMap`
       transformResult.functionMap ??
       null,
+    reactClientReference: transformResult.metadata?.reactClientReference,
   };
 
   return await transformJS(jsFile, context);
@@ -498,7 +515,7 @@ async function transformJSON(
     jsType = 'js/module';
   }
 
-  const output: JsOutput[] = [
+  const output: ExpoJsOutput[] = [
     {
       data: { code, lineCount: countLines(code), map, functionMap: null },
       type: jsType,