+++ packages/expo-video/android/src/main/java/expo/modules/video/VideoPlayer.kt
@@ -50,14 +50,13 @@ class VideoPlayer(context: Context, appContext: AppContext, source: VideoSource?
       field = value
     }
 
-  // This is used only for sending events and keeping the reference to the video source for the
-  // VideoManager, which holds weak references. Changing this will not affect the player.
-  var videoSource: VideoSource? = source
-    set(videoSource) {
-      if (field != videoSource) {
-        sendEventOnJSThread("sourceChange", videoSource, field)
+  var uncommittedSource: VideoSource? = source
+  private var lastLoadedSource: VideoSource? = null
+    set(value) {
+      if (field != value && value != null) {
+        sendEventOnJSThread("sourceChange", value, field)
       }
-      field = videoSource
+      field = value
     }
 
   // Volume of the player if there was no mute applied.
@@ -123,8 +122,6 @@ class VideoPlayer(context: Context, appContext: AppContext, source: VideoSource?
     }
 
     override fun onMediaItemTransition(mediaItem: MediaItem?, reason: Int) {
-      val newVideoSource = VideoManager.getVideoSourceFromMediaItem(mediaItem)
-      this@VideoPlayer.videoSource = newVideoSource
       this@VideoPlayer.duration = 0f
       this@VideoPlayer.isLive = false
       if (reason == Player.MEDIA_ITEM_TRANSITION_REASON_REPEAT) {
@@ -222,7 +219,8 @@ class VideoPlayer(context: Context, appContext: AppContext, source: VideoSource?
       player.removeListener(playerListener)
       player.release()
     }
-    videoSource = null
+    uncommittedSource = null
+    lastLoadedSource = null
   }
 
   override fun deallocate() {
@@ -237,11 +235,12 @@ class VideoPlayer(context: Context, appContext: AppContext, source: VideoSource?
   }
 
   fun prepare() {
-    videoSource?.let { videoSource ->
+    uncommittedSource?.let { videoSource ->
       val mediaItem = videoSource.toMediaItem()
-      VideoManager.registerVideoSourceToMediaItem(mediaItem, videoSource)
       player.setMediaItem(mediaItem)
       player.prepare()
+      lastLoadedSource = videoSource
+      uncommittedSource = null
     } ?: run {
       player.removeMediaItem(0)
       player.prepare()